Move builder API into `monument::query::builder`
